Make "components" folder; move components there
