+++ frontend/src/components/Home.js
@@ -6,9 +6,8 @@ import Typography from '@material-ui/core/Typography';
 import Button from '@material-ui/core/Button';
 import IconButton from '@material-ui/core/IconButton';
 import MenuIcon from '@material-ui/icons/Menu';
-import ButtonAppBar from './ButtonAppBar';
 
-import Signup from './Signup';
+import Signup from '../components/Signup';
 
 import React, { Component } from 'react'
 
@@ -16,8 +15,6 @@ export default class Home extends Component {
   render() {
     return (
       <div>
-        <ButtonAppBar />
-        <br />
         <Signup></Signup>
       </div>
     )
